// src/models/suppliers.ts
export class suppliers{
  supplierid: number;
  firstname: string;
  lastname: string;
  email: string;
  phone: string;
  address: string;
  suppliercode: string;
  status: string;
  bookings: any;
  dprnumber: any;
  company_name: any;
  date_created: Date;
  SuppplierCompanyName: any;
  SuppplierPrice: any;
  orderid: any;
  paycharges:any;
  SupplierDeviceToken:any;
  count: any;
}